backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
pyyaml==6.0.2
watchfiles==1.0.3
httpx==0.28.1
feedparser==6.0.11
cachetools==5.5.0
